+++ server/src/main/scripts/trumpet.sh
@@ -54,7 +54,7 @@ cd $HOMEDIR
 ## Generate classpath from libraries in $LIBDIR, and add $CONFDIR if required
 OUR_CLASSPATH=$(find $LIBDIR -type f -name "*.jar" | paste -sd:)
 if [ -d "$CONFDIR" ]; then
-    OUR_CLASSPATH=$OUR_CLASSPATH:$CONFDIR
+    OUR_CLASSPATH=$CONFDIR:$OUR_CLASSPATH
 fi
 
 # Remote Debug Java Opts
@@ -71,6 +71,7 @@ MX=2048m
 
 JAVA_OPTS="$JAVA_OPTS -Xms$MS -Xmx$MX -XX:+UseParNewGC -XX:+UseConcMarkSweepGC -XX:-CMSConcurrentMTEnabled -XX:CMSInitiatingOccupancyFraction=70 -XX:+CMSParallelRemarkEnabled -XX:+DoEscapeAnalysis"
 JAVA_OPTS="$JAVA_OPTS -Dlogback.configurationFile=logback-production.xml"
+#JAVA_OPTS="$JAVA_OPTS -Dlog4j.configuration=log4j-production.properties"
 
 export HADOOP_USER_CLASSPATH_FIRST=true